FileSys Thread Local, Carriage Return, dirCtx rename

+++ ExtLib.h
@@ -196,7 +196,7 @@ void MemFile_Malloc(MemFile* memFile, u32 size);
 void MemFile_Realloc(MemFile* memFile, u32 size);
 void MemFile_Rewind(MemFile* memFile);
 s32 MemFile_Write(MemFile* dest, void* src, u32 size);
-s32 MemFile_Insert(MemFile* mem, void* src, s32 size, s64 pos);
+s32 MemFile_Insert(MemFile* mem, void* src, u32 size, u32 pos);
 s32 MemFile_Append(MemFile* dest, MemFile* src);
 void MemFile_Align(MemFile* src, u32 align);
 s32 MemFile_Printf(MemFile* dest, const char* fmt, ...);
